Extract publish container context; move ContentDirection into Publish
The ui/index.js export-ordering hack existed because content-direction.js
imported injectContainerContext from the heavy Publish/Container.vue,
so evaluating the ContentDirection export early in the barrel dragged
Container.vue's whole component graph in mid-evaluation and crashed on
a circular-import TDZ (Cannot access 'Dropdown' before initialization).

- Extract the context tuple into Publish/context.js; Container.vue now
  imports and re-exports it for the existing direct importers.
- content-direction.js imports injectContainerContext from the new
  light context.js module instead of Container.vue.
- Move ContentDirection.vue into Publish/, since it's conceptually
  part of the publish container's public surface.
- ui/index.js no longer needs the ordering workaround; ContentDirection
  is now exported from its normal alphabetical position.
