+++ resources/js/components/ui/Publish/ContentDirection.vue
