+++ resources/js/components/ui/Publish/context.js
@@ -1,0 +1,3 @@
+import createContext from '@/util/createContext.js';
+
+export const [injectContainerContext, provideContainerContext, containerContextKey] = createContext('PublishContainer');
